refactor: move useDisconnectWallet into wallet lib
