+++ libs/wallet/src/web3-react/hooks/useDisconnectWallet.ts
@@ -1,0 +1,16 @@
+import { Command } from '@cowprotocol/types'
+import { useWeb3React } from '@web3-react/core'
+
+export function useDisconnectWallet(onDisconnect?: Command) {
+  const { connector } = useWeb3React()
+
+  return () => {
+    if (connector.deactivate) {
+      connector.deactivate()
+    } else {
+      connector.resetState()
+    }
+
+    onDisconnect?.()
+  }
+}
